pyproject.toml
[project]
name = "dupfinder"
version = "1.0.0"
description = "Find duplicate images and videos based on content similarity"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Andreas Sagen"}
]
keywords = ["duplicate", "image", "video", "perceptual-hash", "deduplication"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Video",
    "Topic :: Utilities",
]

dependencies = [
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "ImageHash>=4.3.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
# GPU acceleration options
cuda = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
cupy = [
    "cupy-cuda12x>=12.0.0",
]
opencl = [
    "pyopencl>=2023.1",
]
# All GPU options
gpu = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]

[project.scripts]
# CLI entry points
dupfinder = "dupfinder.cli:main"
find-duplicate-images = "dupfinder.cli:find_images_main"
find-duplicate-videos = "dupfinder.cli:find_videos_main"

[project.urls]
Homepage = "https://github.com/yourusername/dupfinder"
Repository = "https://github.com/yourusername/dupfinder"
Issues = "https://github.com/yourusername/dupfinder/issues"

[build-system]
requires = ["uv_build >= 0.9.11, <0.10.0"]
build-backend = "uv_build"

# UV configuration
[tool.uv]
package = true

[tool.uv.build-backend]
module-name = "dupfinder"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pyright>=1.0.0",
    "sphinx>=7.0.0",
    "furo>=2024.1.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "sphinx-autobuild>=2024.0.0",

]

# Ruff configuration
[tool.ruff]
respect-gitignore = true
line-length = 79
indent-width = 4

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 79
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
section-order = [
  "future",
  "standard-library",
  "first-party",
  "local-folder",
  "third-party",
]

# Mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

# PyTest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/dupfinder"]
branch = true
omit = ["*/tests/*", "*/__init__.py", "*/docx_export.py"]

[tool.coverage.report]
exclude_lines = [
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "def main\\(",
]
show_missing = true
fail_under = 90

# PyRight configuration
[tool.pyright]
include = ["src"]
typeCheckingMode = "basic"

reportPossiblyUnboundVariable = "information"